Mixed tests now use a defined set of rules, valid tests continue checking against new rules
